requirements.txt
PyPDF2==3.0.1
Pillow==10.1.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
pdf2image==1.16.3
requests==2.31.0
transformers==4.36.0
torch==2.1.0
numpy>=1.24.0
